Put back vterm -- content scroll-off problem fixed!
It appears that setting the line-spacing to a positive value can mess
with the content display of vterm (although I didn't encounter this
problem in previous commits of vterm). Anyways, since the problem is
fixed upstream, I can happily put back this terminal emulation that I
consider the best ever!